moved Selectable classes into their own namespace, as Model and Snapshot should be selectable
