+++ Assets/Scripts/Model/Model.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using Constants;
 using Helper;
 using JetBrains.Annotations;
+using Selection;
 using Slicing;
 using Snapshots;
 using UnityEngine;
