+++ Assets/Scripts/Snapshots/SnapshotManager.cs
@@ -4,6 +4,7 @@ using Constants;
 using Extensions;
 using Model;
 using Networking;
+using Selection;
 using Slicing;
 using UnityEngine;
